Make some type utilities not specific to the schema code
