+++ experimental/dds/tree2/src/util/index.ts
@@ -67,5 +67,16 @@ export {
 	Assume,
 	assertValidIndex,
 	assertNonNegativeSafeInteger,
+	objectToMap,
 } from "./utils";
 export { ReferenceCountedBase, ReferenceCounted } from "./referenceCounting";
+
+export {
+	AllowOptional,
+	RequiredFields,
+	OptionalFields,
+	_InlineTrick,
+	_RecursiveTrick,
+	FlattenKeys,
+	AllowOptionalNotFlattened,
+} from "./typeUtils";
